+++ src/main/java/nu/itark/frosk/dataset/IndicatorValues.java
@@ -1,13 +1,11 @@
 package nu.itark.frosk.dataset;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class IndicatorValues {
 
 	private String date;
-	private BigDecimal value;
+	private double value;
 	
 }
